Enable Dot-env
Enable Dot-env for hiding key.

In addition, the root directory of source is changed to src/.
As this changed, the webpack config and exporess is changed too.
And name for 'home' is changed to 'intro'.

There is no script for lint, so it is added with prettier.
Prettier make easy fix code styling.
If you want to check lint, just do `npm run lint`.
If you want to fix lint, just do `npm run fix`.

See Also:
 - https://github.com/nomadcoders/wetube
 - https://eslint.org/docs/user-guide/configuring
 - https://prettier.io/docs/en/integrating-with-linters.html
 -